Extract notes and episode artwork to a new handler
