+++ Modules/Server/Sources/PocketCastsServer/Public/Cache/CacheServerHandler.swift
@@ -8,113 +8,24 @@ public class CacheServerHandler {
 
     public static let noShowNotesMessage = "Unable to find show notes for this episode."
 
-    private let showNotesUrlCache: URLCache
     private let colorsUrlsCache: URLCache
 
-    private var requestingNotes = false
-
-    private var showNotesCompletionBlocks: [(ShowNotesPodcast?) -> Void] = []
-    private var showNotesCachedCompletionBlocks: [(ShowNotesPodcast?) -> Void] = []
+    private lazy var episodeInfoHandler = EpisodeInfoHandler.shared
 
     private init() {
-        showNotesUrlCache = URLCache(memoryCapacity: 1.megabytes, diskCapacity: 10.megabytes, diskPath: "show_notes")
         colorsUrlsCache = URLCache(memoryCapacity: 400.kilobytes, diskCapacity: 5.megabytes, diskPath: "colors")
     }
 
     // MARK: - Show Notes
 
-    struct ShowNotes: Decodable {
-        let podcast: ShowNotesPodcast
-    }
-
-    struct ShowNotesPodcast: Decodable {
-        let episodes: [ShowNotesEpisode]
-    }
-
-    struct ShowNotesEpisode: Decodable {
-        let uuid: String
-        let showNotes: String
-        let image: String?
-    }
-
     public func loadShowNotes(podcastUuid: String, episodeUuid: String, cached: ((String) -> Void)? = nil, completion: ((String?) -> Void)?) {
-        var cachedNotes = ""
-        var didSendCachedNotes = false
-
-        requestShowNotes(for: podcastUuid, cached: { showNotes in
-            if let notes = showNotes?.episodes.first(where: { $0.uuid == episodeUuid })?.showNotes {
-                cached?(notes)
-                cachedNotes = notes
-                didSendCachedNotes = true
-            }
-        }, completion: { showNotes in
-            if let episodeNotes = showNotes?.episodes.first(where: { $0.uuid == episodeUuid })?.showNotes {
-                if didSendCachedNotes, episodeNotes == cachedNotes {
-                    return
-                }
-
-                completion?(episodeNotes)
-            } else if !didSendCachedNotes {
-                completion?(CacheServerHandler.noShowNotesMessage)
-            }
-        })
-    }
-
-    private func requestShowNotes(for podcastUuid: String, cached: @escaping (ShowNotesPodcast?) -> Void, completion: @escaping (ShowNotesPodcast?) -> Void) {
-        showNotesCachedCompletionBlocks.append(cached)
-        showNotesCompletionBlocks.append(completion)
-
-        guard !requestingNotes else { return }
-
-        requestingNotes = true
-
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-
-        let url = ServerHelper.asUrl(ServerConstants.Urls.cache() + "mobile/show_notes/full/\(podcastUuid)")
-        var request = URLRequest(url: url)
-        request.cachePolicy = .reloadRevalidatingCacheData
-
-        if let cachedResponse = showNotesUrlCache.cachedResponse(for: request),
-           let showNotes = try? decoder.decode(ShowNotes.self, from: cachedResponse.data) {
-
-            showNotesCachedCompletionBlocks.forEach { $0(showNotes.podcast) }
-            showNotesCachedCompletionBlocks = []
-            requestingNotes = false
-        }
-
-        URLSession.shared.dataTask(with: request) { [weak self] data, response, _ in
-            if let data = data,
-               let response = response,
-               let showNotes = try? decoder.decode(ShowNotes.self, from: data) {
-                let responseToCache = CachedURLResponse(response: response, data: data)
-                self?.showNotesUrlCache.storeCachedResponse(responseToCache, for: request)
-
-                self?.showNotesCompletionBlocks.forEach { $0(showNotes.podcast) }
-                self?.showNotesCompletionBlocks = []
-            } else {
-                completion(nil)
-            }
-
-            self?.requestingNotes = false
-        }.resume()
+        episodeInfoHandler.loadShowNotes(podcastUuid: podcastUuid, episodeUuid: episodeUuid, cached: cached, completion: completion)
     }
 
     // MARK: - Episode Artwork
 
     public func loadEpisodeArtworkUrl(podcastUuid: String, episodeUuid: String, completion: ((String?) -> Void)?) {
-        var retrievedImage = ""
-        let completionBlock: (ShowNotesPodcast?) -> Void = { episodeNotes in
-            guard let image = episodeNotes?.episodes.first(where: { $0.uuid == episodeUuid })?.image,
-                  image != retrievedImage else {
-                return
-            }
-
-            retrievedImage = image
-            completion?(image)
-        }
-
-        requestShowNotes(for: podcastUuid, cached: completionBlock, completion: completionBlock)
+        episodeInfoHandler.loadEpisodeArtworkUrl(podcastUuid: podcastUuid, episodeUuid: episodeUuid, completion: completion)
     }
 
     public func loadPodcastColors(podcastUuid: String, allowCachedVersion: Bool, completion: @escaping ((String?, String?, String?) -> Void)) {
